Added docker compose file to get redis, elasticsearch and rabbitmq services running.

+++ src/main/java/io/vicarius/assignment/user/UserServiceImpl.java
@@ -1,5 +1,6 @@
 package io.vicarius.assignment.user;
 
+import io.vicarius.assignment.util.Util;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -12,11 +13,17 @@ import java.util.stream.Collectors;
 
 @Service
 public class UserService {
-
     @Autowired
     UserRepository userRepository;
+    @Autowired
+    UserElasticRepository userElasticRepository;
+    @Autowired
+    Util util;
 
-    public UserDTO create(UserDTO userDTO){
+    public UserDTO create(UserDTO userDTO) {
+        //Only printing functions during nighttime
+        if(!util.isDayTime())
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, UserMessages.DB_CHANGES_NOT_ALLOWED_AT_NIGHT);
         if(!StringUtils.hasText(userDTO.getFirstName()))
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, UserMessages.MISSING_FIRST_NAME);
         if(!StringUtils.hasText(userDTO.getFirstName()))
@@ -26,24 +33,43 @@ public class UserService {
 
         UserEntity userEntity = userRepository.save(new UserEntity(userDTO));
         userDTO.setId(userEntity.getId());
+        util.sendCreateOperationToQueue(userDTO);
         return userDTO;
     }
 
     public List<UserDTO> retrieve(){
-        return userRepository.findAll()
-                .stream()
-                .map(UserDTO::new)
-                .collect(Collectors.toList());
+        if(util.isDayTime()){
+            return userRepository.findAll()
+                    .stream()
+                    .map(UserDTO::new)
+                    .collect(Collectors.toList());
+        }else{
+            util.checkDatabaseSincronization();
+            return userElasticRepository.findAll()
+                    .stream()
+                    .map(UserDTO::new)
+                    .collect(Collectors.toList());
+        }
     }
 
     public UserDTO retrieveById(String id){
-        Optional<UserEntity> userEntity = userRepository.findById(id);
-        if(userEntity.isEmpty())
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, UserMessages.USER_NOT_FOUND);
-        return new UserDTO(userEntity.get());
+        if(util.isDayTime()){
+            Optional<UserEntity> userEntity = userRepository.findById(id);
+            if(userEntity.isEmpty())
+                throw new ResponseStatusException(HttpStatus.NOT_FOUND, UserMessages.USER_NOT_FOUND);
+            return new UserDTO(userEntity.get());
+        }else{
+            util.checkDatabaseSincronization();
+            Optional<UserDocument> userDocument = userElasticRepository.findById(id);
+            if(userDocument.isEmpty())
+                throw new ResponseStatusException(HttpStatus.NOT_FOUND, UserMessages.USER_NOT_FOUND);
+            return new UserDTO(userDocument.get());
+        }
     }
 
     public UserDTO update(UserDTO userDTO, String id){
+        if(!util.isDayTime())
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, UserMessages.DB_CHANGES_NOT_ALLOWED_AT_NIGHT);
         if(!StringUtils.hasText(userDTO.getFirstName()))
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, UserMessages.MISSING_FIRST_NAME);
         if(!StringUtils.hasText(userDTO.getFirstName()))
@@ -55,12 +81,16 @@ public class UserService {
             throw new ResponseStatusException(HttpStatus.FORBIDDEN, UserMessages.USER_ALREADY_EXISTS);
         userDTO.setId(id);
         userRepository.save(new UserEntity(userDTO));
+        util.sendUpdateOperationToQueue(userDTO);
         return userDTO;
     }
 
     public void delete(String id){
+        if(!util.isDayTime())
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, UserMessages.DB_CHANGES_NOT_ALLOWED_AT_NIGHT);
         if(userRepository.findById(id).isEmpty())
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, UserMessages.USER_NOT_FOUND);
         userRepository.deleteById(id);
+        util.sendUpdateOperationToQueue(new UserDTO(id));
     }
 }
